Splitted treemap series into serveral files.
